+++ src/pydiverse/pipedag/backend/table/sql/sql.py
@@ -5,11 +5,10 @@ import textwrap
 import warnings
 from collections.abc import Iterable
 from contextlib import contextmanager
-from typing import Any
+from typing import Any, TypeVar
 
 import sqlalchemy as sa
 import sqlalchemy.exc
-import sqlalchemy.sql.elements
 
 from pydiverse.pipedag import Stage, Table
 from pydiverse.pipedag.backend.table.base import BaseTableStore
@@ -35,13 +34,20 @@ from pydiverse.pipedag.context.run_context import DeferredTableStoreOp
 from pydiverse.pipedag.errors import CacheError
 from pydiverse.pipedag.materialize.container import RawSql
 from pydiverse.pipedag.materialize.core import MaterializingTask
-from pydiverse.pipedag.materialize.details import resolve_materialization_details_label
+from pydiverse.pipedag.materialize.details import (
+    BaseMaterializationDetails,
+    resolve_materialization_details_label,
+)
 from pydiverse.pipedag.materialize.metadata import (
     LazyTableMetadata,
     RawSqlMetadata,
     TaskMetadata,
 )
 from pydiverse.pipedag.util.hashing import stable_hash
+
+_materialization_details_subclass = TypeVar(
+    "_materialization_details_subclass", bound=BaseMaterializationDetails
+)
 
 
 class SQLTableStore(BaseTableStore):
@@ -151,6 +157,16 @@ class SQLTableStore(BaseTableStore):
             - a table references a ``materialization_details`` tag that is not defined
               in the config.
         If ``False``: Log an error instead of raising an exception
+
+    :param materialization_details:
+        A dictionary with each entry describing a tag for materialization details of
+        the table store. See subclasses of :py:class:`BaseMaterializationDetails
+         <pydiverse.pipedag.materialize.details.BaseMaterializationDetails>`
+        for details.
+    :param default_materialization_details:
+        The materialization_details that will be used if materialization_details
+        is not specified on table level. If not set, the ``__any__`` tag (if specified)
+        will be used.
     """
 
     METADATA_SCHEMA = "pipedag_metadata"
@@ -158,6 +174,7 @@ class SQLTableStore(BaseTableStore):
 
     __registered_dialects: dict[str, type[SQLTableStore]] = {}
     _dialect_name: str
+    _materialization_details_class: _materialization_details_subclass | None = None
 
     def __new__(cls, engine_url: str, *args, **kwargs):
         if cls != SQLTableStore:
@@ -187,21 +204,16 @@ class SQLTableStore(BaseTableStore):
         print_sql: bool = False,
         no_db_locking: bool = True,
         strict_materialization_details: bool = True,
-        **kwargs,
+        materialization_details: dict[str, dict[str | list[str]]] | None = None,
+        default_materialization_details: str | None = None,
     ):
         super().__init__()
 
-        if kwargs:
-            optional_args = {
-                "materialization_details",
-                "default_materialization_details",
-            }
-            if not set(kwargs.keys()).issubset(optional_args):
-                raise TypeError(
-                    f"The given arguments {set(kwargs.keys()) - optional_args}"
-                    f" are not supported."
-                )
-            error_msg = f"The table store does not support {set(kwargs.keys())}."
+        if (
+            materialization_details is not None
+            or default_materialization_details is not None
+        ) and self._materialization_details_class is None:
+            error_msg = "The table store does not support materialization details."
             if strict_materialization_details:
                 raise TypeError(
                     f"{error_msg} To suppress this exception, "
@@ -290,6 +302,17 @@ class SQLTableStore(BaseTableStore):
             sa.Column("task_hash", sa.String(20)),
             sa.Column("in_transaction_schema", sa.Boolean()),
         )
+
+        if self._materialization_details_class is not None:
+            self.materialization_details = (
+                self._materialization_details_class.create_materialization_details_dict(
+                    materialization_details,
+                    self.strict_materialization_details,
+                    default_materialization_details,
+                    self.logger,
+                )
+            )
+            self.default_materialization_details = default_materialization_details
 
         self.logger.info(
             "Initialized SQL Table Store",
@@ -408,9 +431,14 @@ class SQLTableStore(BaseTableStore):
         return self._execute(query, conn)
 
     def check_materialization_details_supported(self, label: str | None) -> None:
-        from pydiverse.pipedag.backend.table.sql.dialects import IBMDB2TableStore
+        from pydiverse.pipedag.backend.table.sql.dialects import (
+            IBMDB2TableStore,
+            PostgresTableStore,
+        )
 
-        if label is not None and not isinstance(self, IBMDB2TableStore):
+        if label is not None and not isinstance(
+            self, (IBMDB2TableStore, PostgresTableStore)
+        ):
             error_msg = (
                 f"Materialization details are not supported"
                 f" for store {type(self).__name__}."